requirements.txt
google-genai>=1.0.0
chromadb>=0.5.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0
python-multipart>=0.0.9
numpy>=1.26.0
Pillow>=10.0.0
pydantic>=2.0.0
